# specs/002-workflow-engine/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Trámita — Workflow Engine API (motor de trámites)
  version: 0.1.0
  description: |
    Contrato REST de la feature 002 (motor de workflow configurable + timeline de
    auditoría). Toda operación exige la sesión autenticada de la feature 001 (cookie de
    sesión HttpOnly; FR-012): sin sesión válida el backend responde 401 problem+json.
    Las mutaciones exigen además el header X-XSRF-TOKEN espejando la cookie XSRF-TOKEN
    (patrón CSRF de 001). Errores en formato RFC 9457 (application/problem+json).
servers:
  - url: /api
paths:
...
  /requests:
    post:
      summary: Registra una solicitud de trámite (US1)
      description: |
        La solicitud nace en el estado inicial de la definición vigente de su trámite y
        escribe la primera entrada del timeline (autor = sesión actual, from_state null).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateRequestBody' }
      responses:
        '201':
          description: Solicitud creada
          headers:
            Location:
              description: URL de la solicitud creada (/api/requests/{id})
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Request' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422':
          description: Tipo de trámite inexistente en la configuración, o datos inválidos
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
    get:
      summary: Localiza solicitudes por nombre o cédula del estudiante (US3, FR-011)
      parameters:
        - name: search
          in: query
          required: true
          description: |
            Cédula (igualdad exacta) o fragmento del nombre (búsqueda case-insensitive).
            Un solo parámetro para los dos criterios: son los dos datos con los que la
            Coordinación identifica el trámite al recibir la consulta.
          schema: { type: string, minLength: 2 }
      responses:
        '200':
          description: Solicitudes que coinciden (lista vacía si ninguna)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/RequestSummary' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /requests/{id}:
    get:
      summary: Detalle de una solicitud con sus transiciones disponibles
      description: |
        availableTransitions se deriva de la definición con la que nació la solicitud
        (FR-009): la UI se configura sola desde la definición, sin conocer trámites.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Detalle de la solicitud
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Request' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      summary: Corrige una solicitud devuelta o rechazada
      description: |
        Actualiza los datos del formulario sin alterar el timeline. Solo está permitido
        cuando el estado de la solicitud representa una devolución o rechazo; el backend
        vuelve a ejecutar las reglas de negocio configuradas antes de persistir.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateRequestBody' }
      responses:
        '200':
          description: Solicitud corregida
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Request' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: La solicitud no está devuelta ni rechazada
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '422':
          description: Los datos corregidos incumplen una regla del trámite
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /requests/{id}/transitions:
    post:
      summary: Avanza (o devuelve) una solicitud por una transición definida (US2, US5)
      description: |
        Aplica la transición del estado actual hacia targetStateCode si la definición de
        la solicitud la contempla. note es obligatoria cuando la transición la exige
        (devoluciones — FR-014). El motor no distingue avance de retorno (FR-013).
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdvanceRequestBody' }
      responses:
        '200':
          description: Transición aplicada; se devuelve la solicitud actualizada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Request' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: |
            Transición no definida desde el estado actual, estado final (trámite
            cerrado), o conflicto de concurrencia (otro avance ganó — reintentar
            leyendo el estado vigente). El estado no cambia y no se escribe timeline.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '422':
          description: La transición exige observación (devolución) y no llegó (FR-014)
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /requests/{id}/timeline:
    get:
      summary: Timeline de auditoría completo de una solicitud (US3, FR-008)
      description: |
        Todas las entradas en orden cronológico, cada una con autor real, responsable
        del paso según la definición (FR-006) y observación si la hubo. No existe
        operación de edición ni borrado de entradas (FR-007).
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Timeline en orden cronológico ascendente
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TimelineEntry' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /requests/{id}/document:
    get:
      summary: Descarga la constancia PDF de una solicitud finalizada
      description: El documento se genera desde los datos persistidos y solo está disponible al cerrar el trámite.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Documento PDF formal
          content:
            application/pdf: {}
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: La solicitud todavía no está finalizada
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
...
components:
  responses:
    BadRequest:
      description: |
        Cuerpo ausente, JSON malformado, tipo equivocado o violación de bean validation.

        RECTIFICADO EL 2026-09-19. Esta respuesta declaraba que «el detail es genérico
        ("Invalid request content.") y no nombra el campo», porque la emitía el manejador
        heredado de ResponseEntityExceptionHandler. Era cierto y dejó de serlo: se midió que
        TRES causas distintas —campo ausente, campo en blanco y valor fuera de rango—
        colapsaban en esa misma frase en inglés, y que el cliente terminaba deduciendo el
        campo leyendo el código del servidor. Ocurrió: el formulario de novedad de notas
        enviaba `credits = 0` y el trámite era irradicable sin que el error dijera por qué.

        Ahora la violación de bean validation responde en español y nombra los campos, con
        los mismos dos miembros de extensión que el canal público (`missingFields`,
        `invalidFields`). El código NO cambia: sigue siendo 400, porque el 422 del canal
        público está justificado por quién envía allí, no por la clase de error.

        Las demás causas —cuerpo ausente, JSON malformado, tipo equivocado— siguen saliendo
        del manejador heredado con su detail genérico: no hay campo que nombrar cuando el
        cuerpo no se pudo leer.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ValidationProblem' }
    Unauthorized:
      description: Sin sesión autenticada de la Coordinación (FR-012)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: La solicitud no existe
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    ValidationProblem:
      description: |
        `Problem` (RFC 9457) con los dos miembros de extensión que nombran los campos.

        POR QUÉ EXISTEN. El cliente necesita atar el error a SU campo. Sin estos arreglos esa
        información solo viaja dentro del `detail`, una frase en español, y extraerla obliga a
        parsear texto: cualquier retoque de redacción rompería al cliente.

        Los dos llegan SIEMPRE cuando la causa es bean validation, aunque vengan vacíos: una
        forma estable es más barata de consumir que una que aparece y desaparece. Cuando la
        causa es otra —JSON ilegible— la respuesta es un `Problem` sin ellos.

        Se nombran campos, NUNCA valores: el cuerpo rechazado puede traer datos personales y
        el problem+json viaja a la pantalla y a los registros de acceso (§III de la
        constitución, minimización de datos personales).
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          required: [missingFields, invalidFields]
          properties:
            missingFields:
              type: array
              description: |
                Campos que no llegaron diligenciados: ausentes, vacíos o de solo espacios.
                Si hay al menos uno, el `title` es «Petición incompleta».
              items:
                type: string
              example: [studentName]
            invalidFields:
              type: array
              description: |
                Campos que SÍ llegaron pero cuyo valor está fuera de contrato. Un campo que
                incumple las dos cosas a la vez aparece SOLO en `missingFields`: la ausencia
                domina, para no obligar al cliente a decidir cuál mostrar.
              items:
                type: string
              example: ['subjects[0].credits']
    Problem:
      type: object
      description: >-
        RFC 9457 (application/problem+json). Solo title y status son requeridos; detail
        explica el motivo concreto (p. ej. qué transición no está definida).
      required: [title, status]
      properties:
        title: { type: string, example: Conflict }
        status: { type: integer, example: 409 }
        detail: { type: string }
        instance: { type: string }
...
    WorkflowDefinition:
      type: object
      required: [code, name, version]
      properties:
        code: { type: string, example: ADICION_CREDITOS }
        name: { type: string, example: Adición de créditos }
        version: { type: integer, example: 1 }
    CreateRequestBody:
      type: object
      required: [definitionCode, studentName, studentDocument]
      properties:
        definitionCode: { type: string, example: ADICION_CREDITOS }
        studentName: { type: string, maxLength: 120, example: Ana María Pérez }
        studentDocument: { type: string, maxLength: 20, example: '1144099888' }
        studentCode: { type: string, maxLength: 30, example: '1090234' }
        studentEmail: { type: string, format: email, maxLength: 255 }
        program: { type: string, maxLength: 120, example: Ingeniería de Sistemas }
        semester: { type: string, maxLength: 50, example: Semestre 7 }
        reason: { type: string }
        priority: { type: string, enum: [normal, urgente] }
        subjects:
          type: array
          items: { $ref: '#/components/schemas/SubjectRequest' }
    UpdateRequestBody:
      type: object
      required: [studentName, studentDocument]
      properties:
        studentName: { type: string, maxLength: 120 }
        studentDocument: { type: string, maxLength: 20 }
        studentCode: { type: string, maxLength: 30 }
        program: { type: string, maxLength: 120 }
        semester: { type: string, maxLength: 50 }
        reason: { type: string, maxLength: 2000 }
        subjects:
          type: array
          items: { $ref: '#/components/schemas/SubjectRequest' }
    AdvanceRequestBody:
      type: object
      required: [targetStateCode]
      properties:
        targetStateCode: { type: string, example: EN_FACULTAD }
        note:
          type: string
          description: Observación; obligatoria si la transición la exige (FR-014)
    SubjectRequest:
      type: object
      required: [code, name]
      properties:
        code: { type: string, maxLength: 30 }
        name: { type: string, maxLength: 150 }
        credits: { type: integer, nullable: true }
        group: { type: string, maxLength: 30, nullable: true }
        currentGrade: { type: string, maxLength: 20, nullable: true }
        proposedGrade: { type: string, maxLength: 20, nullable: true }
    SubjectResponse:
      type: object
      required: [code, name]
      properties:
        code: { type: string }
        name: { type: string }
        credits: { type: integer, nullable: true }
        group: { type: string, nullable: true }
        currentGrade: { type: string, nullable: true }
        proposedGrade: { type: string, nullable: true }
...
    State:
      type: object
      required: [code, name, isFinal]
      properties:
        code: { type: string, example: EN_FACULTAD }
        name: { type: string, example: En facultad }
        isFinal: { type: boolean }
    AvailableTransition:
      type: object
      required: [targetState, responsible, requiresNote]
      properties:
        targetState: { $ref: '#/components/schemas/State' }
        responsible:
          type: string
          description: Responsable del paso según la definición (FR-006)
          example: FACULTAD
        requiresNote: { type: boolean }
    Request:
      type: object
      required: [id, definition, studentName, studentDocument, currentState,
                 availableTransitions, createdAt]
      properties:
        id: { type: string, format: uuid }
        definition: { $ref: '#/components/schemas/WorkflowDefinition' }
        studentName: { type: string }
        studentDocument: { type: string }
        studentCode: { type: string, nullable: true }
        studentEmail: { type: string, nullable: true }
        program: { type: string, nullable: true }
        semester: { type: string, nullable: true }
        reason: { type: string, nullable: true }
        priority: { type: string, enum: [normal, urgente] }
        subjects:
          type: array
          items: { $ref: '#/components/schemas/SubjectResponse' }
        currentState: { $ref: '#/components/schemas/State' }
        availableTransitions:
          type: array
          description: Vacía cuando el estado actual es final (trámite cerrado)
          items: { $ref: '#/components/schemas/AvailableTransition' }
        createdAt: { type: string, format: date-time }
    RequestSummary:
      type: object
      required: [id, definition, studentName, studentDocument, currentState, createdAt]
      properties:
        id: { type: string, format: uuid }
        definition: { $ref: '#/components/schemas/WorkflowDefinition' }
        studentName: { type: string }
        studentDocument: { type: string }
        studentCode: { type: string, nullable: true }
        studentEmail: { type: string, nullable: true }
        program: { type: string, nullable: true }
        semester: { type: string, nullable: true }
        reason: { type: string, nullable: true }
        priority: { type: string, enum: [normal, urgente] }
        subjects:
          type: array
          items: { $ref: '#/components/schemas/SubjectResponse' }
        currentState: { $ref: '#/components/schemas/State' }
        createdAt: { type: string, format: date-time }
    TimelineEntry:
      type: object
      required: [id, toState, actorEmail, occurredAt]
      properties:
        id: { type: integer, format: int64 }
        fromState:
          oneOf:
            - { $ref: '#/components/schemas/State' }
            - { type: 'null' }
          description: 'null en la entrada de registro (nacimiento de la solicitud)'
        toState: { $ref: '#/components/schemas/State' }
        actorEmail:
          type: string
          description: >-
            Quién registró la entrada — siempre un usuario autenticado de 001, cuyo
            identificador público es el email (users no persiste nombre para mostrar)
        responsible:
          type: string
          description: >-
            En nombre de quién ocurrió el paso, según la definición (FR-006). Ausente en
            la entrada de registro.
        note: { type: string }
        occurredAt: { type: string, format: date-time }